clean up interfaces and tests
