--- Alfonso/Project.toml
@@ -1,2 +1,0 @@
-[deps]
-Alfonso = "b99e6be6-89ff-11e8-14f8-45c827f4f8f2"
