+++ examples/envelope/envelope.jl
@@ -16,12 +16,12 @@ using LinearAlgebra
 using DelimitedFiles
 using Random
 
-function solve_envelope(npoly, deg, n, d; native=true, use_data=false, rseed=1)
+function build_envelope(npoly, deg, n, d; native=true, use_data=false, rseed=1)
     # TODO allow n > 1
     @assert n == 1
 
     # generate interpolation
-    (L, U, pts, P0, P, w) = cheb2_data(d)
+    (L, U, pts, P0, P, w) = Alfonso.cheb2_data(d)
     LWts = fill(binomial(n+d-1, n), n)
     wtVals = 1.0 .- pts.^2
     PWts = [Array((qr(Diagonal(sqrt.(wtVals[:,j]))*P[:,1:LWts[j]])).Q) for j in 1:n]
@@ -39,32 +39,31 @@ function solve_envelope(npoly, deg, n, d; native=true, use_data=false, rseed=1)
         c = vec(P[:,1:LDegs]*rand(-9:9, LDegs, npoly))
     end
 
-    cones = ConeData[]
+    cones = Alfonso.ConeData[]
     coneidxs = AbstractUnitRange[]
     for k in 1:npoly
-        push!(cones, SumOfSqrData(U, P, PWts))
+        push!(cones, Alfonso.SumOfSqrData(U, P, PWts))
         push!(coneidxs, 1+(k-1)*U:k*U)
     end
 
     if native
         # use native interface
         alf = Alfonso.AlfonsoOpt(maxiter=100, verbose=true)
-        Alfonso.loaddata!(alf, A, b, c, cones, coneidxs)
-
-        @time Alfonso.runalgorithm!(alf)
-
-        return (status=alf.status, objval=alf.pobj, objbnd=alf.dobj)
+        Alfonso.load_data!(alf, A, b, c, cones, coneidxs)
+        return alf
     else
         error("MOI tests not implemented")
-        # status = MOI.get(opt, MOI.TerminationStatus())
-        # objval = MOI.get(opt, MOI.ObjectiveValue())
-        # objbnd = MOI.get(opt, MOI.ObjectiveBound())
-        # return (status=status, objval=objval, objbnd=objbnd)
+        # opt = Alfonso.Optimizer()
+        # return opt
     end
 end
 
 # optionally use fixed data in folder
 # select number of polynomials and degrees for the envelope
 # select dimension and SOS degree (to be squared)
-# solve_envelope(2, 5, 1, 5, use_data=true)
-# solve_envelope(2, 5, 1, 5)
+# alf =
+#     build_envelope(2, 5, 1, 5, use_data=true)
+    # build_envelope(2, 5, 1, 5)
+
+# solve it
+# @time Alfonso.solve!(alf)
