+++ examples/lp/lp.jl
@@ -13,7 +13,7 @@ using LinearAlgebra
 using DelimitedFiles
 using Random
 
-function solve_lp(m, n; native=true, use_data=false, dense=false, nzfrac=1/sqrt(n), rseed=1)
+function build_lp(m, n; native=true, use_data=false, dense=false, nzfrac=1/sqrt(n), rseed=1)
     if use_data
         # use provided data in data folder
         datapath = joinpath(@__DIR__, "data")
@@ -32,26 +32,26 @@ function solve_lp(m, n; native=true, use_data=false, dense=false, nzfrac=1/sqrt(
         c = rand(0.0:9.0, n)
     end
 
-    cones = ConeData[NonnegData(n),]
+    cones = Alfonso.ConeData[Alfonso.NonnegData(n),]
     coneidxs = AbstractUnitRange[1:n,]
 
     if native
         # use native interface
         alf = Alfonso.AlfonsoOpt(maxiter=100, verbose=true)
-        Alfonso.loaddata!(alf, A, b, c, cones, coneidxs)
-
-        @time Alfonso.runalgorithm!(alf)
-
-        return (status=alf.status, objval=alf.pobj, objbnd=alf.dobj)
+        Alfonso.load_data!(alf, A, b, c, cones, coneidxs)
+        return alf
     else
         error("MOI tests not implemented")
-        # status = MOI.get(opt, MOI.TerminationStatus())
-        # objval = MOI.get(opt, MOI.ObjectiveValue())
-        # objbnd = MOI.get(opt, MOI.ObjectiveBound())
-        # return (status=status, objval=objval, objbnd=objbnd)
+        # opt = Alfonso.Optimizer()
+        # return opt
     end
 end
 
 # optionally use fixed data in folder
 # select the random matrix size, dense/sparse, sparsity fraction
-# solve_lp(500, 1000)
+# alf =
+    # build_lp(500, 1000)
+    # build_lp(500, 1000, use_data=true)
+
+# solve it
+# @time Alfonso.solve!(alf)
