+++ src/conedata.jl
@@ -1,5 +1,5 @@
 
-export ConeData, NonnegData, SumOfSqrData
+# export ConeData, NonnegData, SumOfSqrData
 
 
 abstract type ConeData end
