+++ test/testnative.jl
@@ -1,40 +1,45 @@
 
 include(joinpath(@__DIR__, "../examples/envelope/envelope.jl"))
 @testset "envelope example" begin
-    r = solve_envelope(2, 5, 1, 5, use_data=true)
-    @test r.status == :Optimal
-    @test r.objval ≈ -25.502777 rtol=1e-4
-    @test r.objbnd ≈ r.objval rtol=1e-4
+    alf = build_envelope(2, 5, 1, 5, use_data=true)
+    Alfonso.solve!(alf)
+    @test Alfonso.get_status(alf) == :Optimal
+    @test Alfonso.get_pobj(alf) ≈ -25.502777 rtol=1e-4
+    @test Alfonso.get_pobj(alf) ≈ -25.502777 rtol=1e-4
 end
 
 include(joinpath(@__DIR__, "../examples/lp/lp.jl"))
 @testset "lp example" begin
-    r = solve_lp(500, 1000, use_data=true)
-    @test r.status == :Optimal
-    @test r.objval ≈ 2055.807 rtol=1e-4
-    @test r.objbnd ≈ r.objval rtol=1e-4
+    alf = build_lp(500, 1000, use_data=true)
+    Alfonso.solve!(alf)
+    @test Alfonso.get_status(alf) == :Optimal
+    @test Alfonso.get_pobj(alf) ≈ 2055.807 rtol=1e-4
+    @test Alfonso.get_pobj(alf) ≈ 2055.807 rtol=1e-4
 end
 
 include(joinpath(@__DIR__, "../examples/namedpoly/namedpoly.jl"))
 @testset "namedpoly examples" begin
     @testset "Goldstein Price" begin
-        r = solve_namedpoly(:goldsteinprice, 7)
-        @test r.status == :Optimal
-        @test r.objval ≈ 3 rtol=1e-4
-        @test r.objbnd ≈ r.objval rtol=1e-4
+        alf = build_namedpoly(:goldsteinprice, 7)
+        Alfonso.solve!(alf)
+        @test Alfonso.get_status(alf) == :Optimal
+        @test Alfonso.get_pobj(alf) ≈ 3 rtol=1e-4
+        @test Alfonso.get_pobj(alf) ≈ 3 rtol=1e-4
     end
 
     @testset "Robinson" begin
-        r = solve_namedpoly(:robinson, 8)
-        @test r.status == :Optimal
-        @test r.objval ≈ 0.814814 rtol=1e-4
-        @test r.objbnd ≈ r.objval rtol=1e-4
+        alf = build_namedpoly(:robinson, 8)
+        Alfonso.solve!(alf)
+        @test Alfonso.get_status(alf) == :Optimal
+        @test Alfonso.get_pobj(alf) ≈ 0.814814 rtol=1e-4
+        @test Alfonso.get_pobj(alf) ≈ 0.814814 rtol=1e-4
     end
 
     @testset "Lotka Volterra" begin
-        r = solve_namedpoly(:lotkavolterra, 3)
-        @test r.status == :Optimal
-        @test r.objval ≈ -20.8 rtol=1e-4
-        @test r.objbnd ≈ r.objval rtol=1e-4
+        alf = build_namedpoly(:lotkavolterra, 3)
+        Alfonso.solve!(alf)
+        @test Alfonso.get_status(alf) == :Optimal
+        @test Alfonso.get_pobj(alf) ≈ -20.8 rtol=1e-4
+        @test Alfonso.get_pobj(alf) ≈ -20.8 rtol=1e-4
     end
 end
